Incremental commit, everything working
