+++ .prettierrc.json
@@ -1,14 +1,7 @@
 {
     "trailingComma": "es5",
     "tabWidth": 4,
-    "semi": true,
-    "singleQuote": false,
-    "printWidth": 160,
-    "useTabs": false,
-    "bracketSpacing": true,
-    "jsxBracketSameLine": false,
-    "arrowParens": "avoid",
-    "plugins": ["prettier-plugin-svelte"],
-    "pluginSearchDirs": ["."],
-    "overrides": [{ "files": "*.svelte", "options": { "parser": "svelte" } }]
+    "printWidth": 120,
+    "semi": false,
+    "singleQuote": true
 }
