+++ templates/calendar.html
@@ -1,31 +1,39 @@
 {{ define "calendar" }}
 <div class="calendar-container">
+    <div class="calendar-nav">
+        <div class="month-prev" hx-post="/calendar" hx-target=".calendar" hx-vals='{"offset": "{{ .offsetPrev }}"}'>
+            Prev
+        </div>
+        {{ if eq .offset 0 }}
+        <div class="month current" hx-post="/calendar" hx-target=".calendar" hx-vals='{"offset": 0}'>
+            {{ .month }} {{ .year }}
+        </div>
+        {{ else }}
+        <div class="month" hx-post="/calendar" hx-target=".calendar" hx-vals='{"offset": 0}'>
+            {{ .month }} {{ .year }}
+        </div>
+        {{ end }}
+        <div class="month-next" hx-post="/calendar" hx-target=".calendar" hx-vals='{"offset": "{{ .offsetNext }}"}'>
+            Next
+        </div>
+    </div>
     <div class="calendar-grid">
         <div class="corner"></div>
         {{ range .weekdays }}
-            <div class="weekday">{{ . }}</div>
+        <div class="weekday">{{ . }}</div>
+        {{ end }} {{ range $key, $value := .weeks }} {{ if and (eq $value $.currentWeek) (eq $.offset 0) }}
+        <div class="week current" style="grid-row: {{$key}};">{{$value}}</div>
+        {{ else }}
+        <div class="week" style="grid-row: {{$key}};">{{$value}}</div>
+        {{ end }} {{ end }} {{ range .prevDays }}
+        <div class="prev-day">{{ . }}</div>
+        {{ end }} {{ range .days }} {{ if and (eq . $.currentDay) (eq $.offset 0) }}
+        <div class="day current">{{ . }}</div>
+        {{ else }}
+        <div class="day">{{ . }}</div>
+        {{ end }} {{ end }} {{ range .nextDays }}
+        <div class="next-day">{{ . }}</div>
         {{ end }}
-        {{ range $key, $value := .weeks }}
-            {{ if eq $value $.week }}
-                <div class="week current" style="grid-row: {{$key}};">{{$value}}</div>
-            {{ else }}
-            <div class="week" style="grid-row: {{$key}};">{{$value}}</div>
-            {{ end }}
-        {{ end }}
-          {{ range .prevDays }}
-            <div class="prev-day">{{ . }}</div>
-        {{ end }}
-        {{ range .days }}
-  {{ if eq . $.day }}
-    <div class="day current">{{ . }}</div>
-  {{ else }}
-    <div class="day">{{ . }}</div>
-  {{ end }}
-{{ end }}
-       {{ range .nextDays }}
-            <div class="next-day">{{ . }}</div>
-        {{ end }}
-
     </div>
 </div>
 {{ end }}
