+++ templates/header.html
@@ -1,3 +1,3 @@
 {{ define "header" }}
-    <p>This is Index body</p>
-{{ end }}
+    <p>I'm just an header.</p>
+{{ end  }}
